refactor(project): extract own-certs quick-pick into useOwnCerts [quality-056-cert-6]
Dedupes the listRecords quick-pick fetch shared by /project/new and project edit.
